Spring version update New versioning Added NardaConnector Correct task updating Correct float values (eb_no_delta and eb_no_remote_delta)

+++ src/main/java/ru/sereda/autowiretest/DTO/DeviceTimerTaskDTO.java
@@ -4,10 +4,10 @@ package ru.sereda.autowiretest.DTO;
 public class DeviceTimerTaskDTO {
     private String taskId;
     private boolean sleep;
-    private int interval;
+    private long interval;
     private String deviceId;
 
-    public DeviceTimerTaskDTO(String taskId, boolean sleep, int interval, String deviceId) {
+    public DeviceTimerTaskDTO(String taskId, boolean sleep, long interval, String deviceId) {
         this.taskId = taskId;
         this.sleep = sleep;
         this.interval = interval;
@@ -23,7 +23,7 @@ public class DeviceTimerTaskDTO {
         return sleep;
     }
 
-    public int getInterval() {
+    public long getInterval() {
         return interval;
     }
